Fields for a Muschelli Package

Fields for a Muschelli Package

muschelli_fields(title = "", description = "", given = "John",
  family = "Muschelli", email = "muschellij2@gmail.com", role = c("aut",
  "cre"))

Arguments

title
Title of package

description
Description of package. Should be able to work with the start of the sentence "The goal is to provide".

given
a character vector with the given names, or a list thereof.

family
a character string with the family name, or a list thereof.

email
a character string (or vector) giving an e-mail address (each), or a list thereof.

role
Role of the people (author/creator)

Value

A list of named fields

Examples

muschelli_fields()

#> $Type
#> [1] "Package"
#> 
#> $Title
#> [1] ""
#> 
#> $Description
#> [1] ""
#> 
#> $Version
#> [1] "0.1.0"
#> 
#> $`Authors@R`
#> [1] "structure(list(structure(list(given = \"John\", family = \"Muschelli\", \n        role = c(\"aut\", \"cre\"), email = \"muschellij2@gmail.com\", \n        comment = NULL), .Names = c(\"given\", \"family\", \"role\", \"email\", \n    \"comment\"))), class = \"person\")"
#> 
#> $Maintainer
#> [1] "John Muschelli <muschellij2@gmail.com>"
#> 
#> $Encoding
#> [1] "UTF-8"
#> 
#> $LazyData
#> [1] "true"
#> 
#> $Suggests
#> [1] "knitr, rmarkdown, covr"
#> 
#> $VignetteBuilder
#> [1] "knitr"
#>